Remove SR_TEST_NUM test constant
We needn't reorder tests by having to change a bunch of numbers, and the
numbers themselves (ie, what order they go in) shouldn't be spread across
a million files. So instead, order will be defined by what order objects
are in the makefile, and we can get rid of the SR_TEST_NUM value.

Still need to keep the SR_TEST_NUM_STR value, as each test needs a unique
identifier